function/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.9
    # via gcsfs
aiosignal==1.3.2
    # via aiohttp
attrs==25.3.0
    # via aiohttp
blinker==1.9.0
    # via flask
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   flask
    #   functions-framework
cloudevents==1.12.0
    # via functions-framework
db-dtypes==1.4.3
    # via
    #   volans-parser (pyproject.toml)
    #   pandas-gbq
decorator==5.2.1
    # via gcsfs
deprecation==2.1.0
    # via cloudevents
et-xmlfile==2.0.0
    # via openpyxl
flask==3.1.1
    # via functions-framework
frozenlist==1.6.2
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via gcsfs
functions-framework==3.8.3
    # via volans-parser (pyproject.toml)
gcsfs==2025.5.1
    # via volans-parser (pyproject.toml)
google-api-core==2.25.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-storage
    #   pandas-gbq
google-auth==2.40.3
    # via
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-storage
    #   pandas-gbq
    #   pydata-google-auth
google-auth-oauthlib==1.2.2
    # via
    #   gcsfs
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-bigquery==3.34.0
    # via
    #   volans-parser (pyproject.toml)
    #   pandas-gbq
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-storage==3.1.0
    # via gcsfs
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.72.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.72.1
    # via google-api-core
gunicorn==23.0.0
    # via functions-framework
idna==3.10
    # via
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
numpy==2.2.6
    # via
    #   db-dtypes
    #   pandas
    #   pandas-gbq
oauthlib==3.2.2
    # via requests-oauthlib
openpyxl==3.1.5
    # via volans-parser (pyproject.toml)
packaging==25.0
    # via
    #   db-dtypes
    #   deprecation
    #   google-cloud-bigquery
    #   gunicorn
    #   pandas-gbq
pandas==2.3.0
    # via
    #   volans-parser (pyproject.toml)
    #   db-dtypes
    #   pandas-gbq
pandas-gbq==0.29.0
    # via volans-parser (pyproject.toml)
pendulum==3.1.0
    # via volans-parser (pyproject.toml)
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via google-api-core
protobuf==6.31.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyarrow==20.0.0
    # via
    #   db-dtypes
    #   pandas-gbq
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydata-google-auth==1.9.1
    # via pandas-gbq
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   pandas
    #   pendulum
pytz==2025.2
    # via pandas
requests==2.32.3
    # via
    #   volans-parser (pyproject.toml)
    #   gcsfs
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
setuptools==80.9.0
    # via
    #   pandas-gbq
    #   pydata-google-auth
six==1.17.0
    # via python-dateutil
tzdata==2025.2
    # via
    #   pandas
    #   pendulum
unidecode==1.4.0
    # via volans-parser (pyproject.toml)
urllib3==2.4.0
    # via requests
watchdog==6.0.0
    # via functions-framework
werkzeug==3.1.3
    # via
    #   flask
    #   functions-framework
xlrd==2.0.1
    # via volans-parser (pyproject.toml)
yarl==1.20.0
    # via aiohttp
